% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_migration.R
\name{db_migrate}
\alias{db_migrate}
\title{Migrate table from source to target}
\usage{
db_migrate(
  source_conn,
  target_conn,
  source_table,
  source_schema = NULL,
  target_schema = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{source_conn}{source db connection}

\item{target_conn}{target db connection}

\item{source_table}{source table name}

\item{source_schema}{source schema}

\item{target_schema}{target schema}
}
\description{
Migrate table from source to target
}
