% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_column_comment.R
\name{drop_column_comment}
\alias{drop_column_comment}
\title{Drop column comment}
\usage{
drop_column_comment(conn, schema = NULL, table, column)
}
\arguments{
\item{conn}{A DBIConnection object.}

\item{schema}{The name of the parent schema of the database table.
If no schema is specified, the default schema of the database is used
(e.g. 'dbo' for Microsoft SQL Server).}

\item{table}{The name of the database table.}

\item{column}{The name of the column.}
}
\description{
This function deletes an existing column comment of a database table.
}
\examples{
delete_column_comment(conn = con, schema = "schema", table = "table", column = "column")

}
