% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_constraint.R
\name{drop_constraint}
\alias{drop_constraint}
\title{Drop constraint}
\usage{
drop_constraint(conn, schema = NULL, table, constraint)
}
\arguments{
\item{conn}{A DBIConnection object.}

\item{schema}{The name of the parent schema of the database table.
If no schema is specified, the default schema of the database is used
(e.g. 'dbo' for Microsoft SQL Server).}

\item{table}{The name of the database table.}

\item{constraint}{The name of the constraint.}
}
\description{
This function removes a specific constraint (e.g. Primary Key, Foreign Key,
Unique, Check) from the specified database table.
}
