% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_table_comment.R
\name{drop_table_comment}
\alias{drop_table_comment}
\title{Drop table comment}
\usage{
drop_table_comment(con, schema = NULL, table)
}
\arguments{
\item{schema}{The name of the parent schema of the database table.
If no schema is specified, the default schema of the database is used
(e.g. 'dbo' for Microsoft SQL Server).}

\item{table}{The name of the database table.}

\item{conn}{A DBIConnection object.}
}
\description{
This function deletes the existing comment of the selected database table.
}
\examples{
drop_table_comment(conn = con, schema = "schema", table = "table")

}
