% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_comment.R
\name{get_column_comment}
\alias{get_column_comment}
\title{Get column comment}
\usage{
get_column_comment(conn, schema = NULL, table, column)
}
\arguments{
\item{conn}{A DBIConnection object.}

\item{schema}{The name of the parent schema of the database table.
If no schema is specified, the default schema of the database is used
(e.g. 'dbo' for Microsoft SQL Server).}

\item{table}{The name of the database table.}

\item{column}{The column name}
}
\description{
This function returns the column comment of a database table.
}
\examples{
get_column_comment(conn = con, schema = "schema", table = "table", column = "column")

}
