% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_database_name.R
\name{get_database_name}
\alias{get_database_name}
\title{Get database name}
\usage{
get_database_name(conn)
}
\arguments{
\item{conn}{A DBIConnection object.}
}
\value{
A character string representing the database name.
}
\description{
This function returns the database name under the used database connection.
}
\examples{
get_database_name(conn = con)

}
