% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_database_type.R
\name{get_database_type}
\alias{get_database_type}
\title{Detect database type}
\usage{
get_database_type(conn)
}
\arguments{
\item{conn}{A DBIConnection object.}
}
\value{
A character string representing the type of the database.
}
\description{
This function returns the detected type of the database connection.
}
\examples{
get_database_type(conn = con)

}
