% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_comment.R
\name{get_table_comment}
\alias{get_table_comment}
\title{Get table comment}
\usage{
get_table_comment(conn, schema = NULL, table)
}
\arguments{
\item{conn}{A DBIConnection object.}

\item{schema}{The name of the parent schema of the database table.
If no schema is specified, the default schema of the database is used
(e.g. 'dbo' for Microsoft SQL Server).}

\item{table}{The name of the database table.}
}
\value{
A character string representing the table comment.
}
\description{
This function returns the comment of the selected database table.
}
\examples{
get_table_comment(conn = con, schema = "schema", table = "table")

}
