% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_permissions.R
\name{get_user_permissions}
\alias{get_user_permissions}
\title{Get permissions of database user}
\usage{
get_user_permissions(conn)
}
\arguments{
\item{conn}{A DBIConnection object.}
}
\value{
A character vector representing the permissions of the database user.
}
\description{
This function returns the permissions of the database user.
}
\examples{
get_user_permissions(conn = con)

}
