% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helper.R
\name{list_databases}
\alias{list_databases}
\title{List Databases}
\usage{
list_databases(conn = NULL)
}
\arguments{
\item{conn}{connection object. Will search global environment for "con" if conn is NULL}

\item{schema}{schema name}
}
\description{
List Databases
}
