% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helper.R
\name{list_tables}
\alias{list_tables}
\title{List Tables}
\usage{
list_tables(conn = NULL, schema = NULL, pattern = ".*")
}
\arguments{
\item{conn}{connection object. Will search global environment for "con" if conn is NULL}

\item{schema}{schema name}
}
\description{
List Tables
}
