% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_users.R
\name{list_users}
\alias{list_users}
\title{List database users}
\usage{
list_users(conn)
}
\arguments{
\item{conn}{A DBIConnection object.}
}
\description{
This function lists all database users and returns them as a dataframe object.
}
