% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_execute.R
\name{run_sql}
\alias{run_sql}
\alias{run_sql_script}
\alias{run_sql_script_by}
\title{Run SQL script from file}
\usage{
run_sql(
  conn,
  sql_statement,
  ...,
  start_date = NA,
  end_date = NA,
  by = NA,
  offset = NA,
  verbose = FALSE
)

run_sql_script(
  conn,
  filename,
  ...,
  start_date = NA,
  end_date = NA,
  by = NA,
  offset = NA,
  verbose = FALSE
)

run_sql_script_by(conn, filename, start_date, end_date, by = "month")
}
\arguments{
\item{conn}{database connection}

\item{sql_statement}{sql statement to run. Multiple sql statements should be seperated by semicolon(;)}

\item{by}{period definition in string}

\item{filename}{file name with full path}
}
\description{
Run SQL script from file
}
