% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_execute.R
\name{run_sql_script_by}
\alias{run_sql_script_by}
\title{Run SQL script by period}
\usage{
run_sql_script_by(conn, filename, start_date, end_date, by = "month")
}
\arguments{
\item{filename}{file name with full path}

\item{start_date}{start date in format YYYY-MM-DD}

\item{end_date}{end date in format YYYY-MM-DD}

\item{by}{period definition in string}
}
\description{
Run SQL script by period
}
