% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_column_comment.R
\name{set_column_comment}
\alias{set_column_comment}
\title{Set column comment}
\usage{
set_column_comment(conn, schema = NULL, table, column, comment)
}
\arguments{
\item{conn}{A DBIConnection object.}

\item{schema}{The name of the parent schema of the database table.
If no schema is specified, the default schema of the database is used
(e.g. 'dbo' for Microsoft SQL Server).}

\item{table}{The name of the database table}

\item{column}{The column name}

\item{comment}{The comment you want to set for this column}
}
\description{
This function sets a comment for the selected column of a database table.
}
\examples{
set_column_comment(conn = con, schema = "schema", table = "table", column = "column", comment = "Your column comment")

}
