% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_primary_key.R
\name{set_primary_key}
\alias{set_primary_key}
\title{Set primary key}
\usage{
set_primary_key(conn, schema = NULL, table, columns, pk_name = NULL)
}
\arguments{
\item{conn}{A DBIConnection object.}

\item{schema}{The name of the parent schema of the database table.
If no schema is specified, the default schema of the database is used
(e.g. 'dbo' for Microsoft SQL Server).}

\item{table}{The name of the database table.}

\item{columns}{Character vector. The column(s) to set as the primary key.}

\item{pk_name}{The optional name of the primary key. If no name is specified,
the primary key name is generated as follows: \cr
\code{pk_name <- paste0("PK_", table)}}
}
\description{
This function sets a primary key for a selected database table. \cr
Note: For Microsoft SQL Server databases, please ensure that the primary key
columns are defined as \code{NOT NULL}.
}
