% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helper.R
\name{tbbl}
\alias{tbbl}
\title{Connect DB Table}
\usage{
tbbl(conn, table_name)
}
\arguments{
\item{conn}{connection object. Will search global environment for "con" if conn is NULL}

\item{table_name}{table name, could be full name with schema}
}
\description{
Connect DB Table
}
