% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_table_comment.R
\name{set_table_comment}
\alias{set_table_comment}
\title{Create a comment for a database table}
\usage{
set_table_comment(conn, schema, table_name, comment)
}
\arguments{
\item{conn}{A DBIConnection object}

\item{schema}{The name of the parent schema of the table}

\item{table_name}{The table name of the specific table}

\item{comment}{The comment you want to use for this table}
}
\description{
Create a comment for a database table
}
